pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-overlay-tool"
version = "0.1.0"
description = "Cross-platform desktop utility for placing one image over another."
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "PySide6>=6.6",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0"]
test = []

[project.scripts]
image-overlay-tool = "image_overlay_tool.cli:main"

[tool.setuptools.packages.find]
include = ["image_overlay_tool*"]
